Wrote simple XML parser for ArXiV dataset.
The parser currently retrieves all <title> and <summary> tags inside
<entry> tags in the ArXiV chunk files. The user specifies an input
directory, output directory, and number of files total to parse (0 for
all files in input directory).

There is currently a small bug wherein the last lines of the outputted
files are the summaries of a entry but cut off early.

Currently, the parsed files take up about 400MB of space and will not be
committed to this repo.